Schema.Truck (FoodTruckApi v0.1.0)

This module is responsible for defining the Truck struct and its functions.

Fields

objectid - The unique identifier for the food truck.

applicant - The name of the person or company that owns the food truck.

facility_type - The type of facility the food truck is.

cnn - The unique identifier for the food truck.

location_description - A description of the location of the food truck.

address - The address of the food truck.

block_lot - The block and lot number of the food truck.

block - The block number of the food truck.

lot - The lot number of the food truck.

permit - The permit number of the food truck.

status - The status of the food truck.

food_items - The food items the food truck serves.

x - The x coordinate of the food truck.

y - The y coordinate of the food truck.

latitude - The latitude of the food truck.

longitude - The longitude of the food truck.

schedule - The schedule of the food truck.

days_hours - The days and hours the food truck is open.

received - The date the food truck received its permit.

prior_permit - Whether the food truck had a prior permit.

location - The location of the food truck.

approved - The date the food truck was approved.

expiration_date - The expiration date of the food truck's permit.

zip_code - The zip code of the food truck.

Example

%FoodTruckApi.Models.Truck{
  objectid: 1,
  applicant: "John Doe",
  facility_type: "Push Cart",
  cnn: 123456,
  location_description: "MARKET ST: DRUMM ST intersection",
  address: "123 Market St",
  block_lot: 234017,
  block: 234,
  lot: 17,
  permit: "15MFF-0159",
  status: "REQUESTED",
  food_items: "Fried Chicken: Fried Fish: Greens: Mac & Cheese: Peach Cobbler: and String beans",
  x: 123.456,
  y: 456.789,
  latitude: 123.456,
  longitude: 456.789,
  schedule: "http://bsm.sfdpw.org/PermitsTracker/reports/report.aspx?title=schedule&report=rptSchedule&params=permit=19MFF-00112&ExportPDF=1&Filename=19MFF-00112_schedule.pdf",
  days_hours: "Mo:6AM-8PM",
  received: 20151231,
  prior_permit: true,
  location: "(37.794331003246846, -122.39581105302317)",
  approved: "03/16/2017 12:00:00 AM",
  expiration_date: "07/15/2018 12:00:00 AM",
  zip_code: 94111
}

Types

t()

@type t() :: %Schema.Truck{
  address: String.t(),
  applicant: String.t(),
  approved: String.t(),
  block: integer(),
  block_lot: String.t(),
  cnn: integer(),
  days_hours: String.t(),
  expiration_date: String.t(),
  facility_type: String.t(),
  food_items: String.t(),
  latitude: float(),
  location: %{
    human_address: String.t(),
    latitude: String.t(),
    longitude: String.t()
  },
  location_description: String.t(),
  longitude: float(),
  lot: integer(),
  objectid: integer(),
  permit: String.t(),
  prior_permit: boolean(),
  received: String.t(),
  schedule: String.t(),
  status: String.t(),
  x: float(),
  y: float(),
  zip_code: integer()
}

Functions

build_data_from_query(data)

@spec build_data_from_query(String.t()) :: [t()]

Converts a JSON object to a Truck struct.

list_food_items(data, food_item)

@spec list_food_items(String.t(), String.t()) :: [t()]

Returns a list of the desired food items.

list_taco_trucks(data)

@spec list_taco_trucks(String.t()) :: [t()]

Returns a list of taco trucks.