FoodTruckApiWeb.TruckController (FoodTruckApi v0.1.0)

Functions

index(conn, params)

@spec index(Plug.Conn.t(), map()) :: [Schema.Truck.t()] | {:error, String.t()}

Retrieves a list of food trucks in the San Francisco area.

Examples

iex> TruckController.index(%Plug.Conn{}, %{})
{:ok, [%Truck{}]}

iex> TruckController.index(%Plug.Conn{}, %{})
{:error, "Not Found"}

iex> TruckController.index(%Plug.Conn{}, %{})
{:error, "Internal Server Error: reason"}

Route

GET /api/trucks

list_approved_trucks(conn, params)

@spec list_approved_trucks(Plug.Conn.t(), map()) ::
  [Schema.Truck.t()] | {:error, String.t()}

Retrieves a list of food trucks with an approved permit in the San Francisco area.

Examples

iex> TruckController.list_approved_trucks(%Plug.Conn{}, %{})
{:ok, [%Truck{}]}

iex> TruckController.list_approved_trucks(%Plug.Conn{}, %{})
{:error, "Not Found"}

iex> TruckController.list_approved_trucks(%Plug.Conn{}, %{})
{:error, "Internal Server Error: reason"}

Route

GET /api/trucks/approved

list_taco_trucks(conn, params)

@spec list_taco_trucks(Plug.Conn.t(), map()) ::
  [Schema.Truck.t()] | {:error, String.t()}

Retrieves a list of taco trucks with an approved permit in the San Francisco area.

Examples

iex> TruckController.list_taco_trucks(%Plug.Conn{}, %{})
{:ok, [%Truck{}]}

iex> TruckController.list_taco_trucks(%Plug.Conn{}, %{})
{:error, "Not Found"}

iex> TruckController.list_taco_trucks(%Plug.Conn{}, %{})
{:error, "Internal Server Error: reason"}

Route

GET /api/trucks/taco